+++ packages/core-components/src/layout/Sidebar/icons/DoubleArrowRight.tsx
@@ -13,10 +13,10 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-
-import React from 'react';
+import Box from '@material-ui/core/Box';
 import { makeStyles } from '@material-ui/core/styles';
 import ArrowForwardIosIcon from '@material-ui/icons/ArrowForwardIos';
+import React from 'react';
 
 const useStyles = makeStyles({
   iconContainer: {
@@ -34,14 +34,14 @@ const DoubleArrowRight = () => {
   const classes = useStyles();
 
   return (
-    <div className={classes.iconContainer}>
-      <div className={classes.arrow1}>
+    <Box className={classes.iconContainer}>
+      <Box className={classes.arrow1}>
         <ArrowForwardIosIcon style={{ fontSize: '12px' }} />
-      </div>
-      <div>
+      </Box>
+      <Box>
         <ArrowForwardIosIcon style={{ fontSize: '12px' }} />
-      </div>
-    </div>
+      </Box>
+    </Box>
   );
 };