+++ plugins/search-backend-module-pg/migrations/20210727180000_delta_update.js
@@ -14,6 +14,8 @@
  * limitations under the License.
  */
 
+// @ts-check
+
 /**
  * @param {import('knex').Knex} knex
  */