+++ packages/backend-app-api/CHANGELOG.md
@@ -1,5 +1,28 @@
 # @backstage/backend-app-api
 
+## 0.2.4-next.2
+
+### Patch Changes
+
+- 884d749b14: Refactored to use `coreServices` from `@backstage/backend-plugin-api`.
+- Updated dependencies
+  - @backstage/backend-common@0.17.0-next.2
+  - @backstage/backend-plugin-api@0.2.0-next.2
+  - @backstage/backend-tasks@0.4.0-next.2
+  - @backstage/plugin-permission-node@0.7.2-next.2
+  - @backstage/errors@1.1.4-next.1
+
+## 0.2.4-next.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/backend-common@0.17.0-next.1
+  - @backstage/backend-tasks@0.4.0-next.1
+  - @backstage/backend-plugin-api@0.1.5-next.1
+  - @backstage/plugin-permission-node@0.7.2-next.1
+  - @backstage/errors@1.1.4-next.1
+
 ## 0.2.4-next.0
 
 ### Patch Changes