+++ plugins/apache-airflow/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @backstage/plugin-apache-airflow
 
+## 0.2.5-next.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/core-plugin-api@1.2.0-next.2
+  - @backstage/core-components@0.12.1-next.2
+
+## 0.2.5-next.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/core-components@0.12.1-next.1
+  - @backstage/core-plugin-api@1.1.1-next.1
+
 ## 0.2.5-next.0
 
 ### Patch Changes