+++ plugins/catalog-graphql/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @backstage/plugin-catalog-graphql
 
+## 0.3.16-next.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/types@1.0.2-next.1
+  - @backstage/config@1.0.5-next.1
+  - @backstage/catalog-model@1.1.4-next.1
+
 ## 0.3.16-next.0
 
 ### Patch Changes